rename dazzler.dash.timeseries to dazzler.dash.fiware.
